+++ lib/internal/bootstrap/switches/does_own_process_state.js
@@ -1,6 +1,12 @@
 'use strict';
 
 const credentials = internalBinding('credentials');
+const rawMethods = internalBinding('process_methods');
+
+process.abort = rawMethods.abort;
+process.umask = wrappedUmask;
+process.chdir = wrappedChdir;
+process.cwd = wrappedCwd;
 
 if (credentials.implementsPosixCredentials) {
   const wrapped = wrapPosixCredentialSetters(credentials);
@@ -15,6 +21,11 @@ if (credentials.implementsPosixCredentials) {
 
 // ---- keep the attachment of the wrappers above so that it's easier to ----
 // ----              compare the setups side-by-side                    -----
+
+const {
+  parseMode,
+  validateString
+} = require('internal/validators');
 
 function wrapPosixCredentialSetters(credentials) {
   const {
@@ -93,4 +104,28 @@ function wrapPosixCredentialSetters(credentials) {
     setgid: wrapIdSetter('Group', _setgid),
     setuid: wrapIdSetter('User', _setuid)
   };
+}
+
+// Cache the working directory to prevent lots of lookups. If the working
+// directory is changed by `chdir`, it'll be updated.
+let cachedCwd = '';
+
+function wrappedChdir(directory) {
+  validateString(directory, 'directory');
+  rawMethods.chdir(directory);
+  // Mark cache that it requires an update.
+  cachedCwd = '';
+}
+
+function wrappedUmask(mask) {
+  if (mask !== undefined) {
+    mask = parseMode(mask, 'mask');
+  }
+  return rawMethods.umask(mask);
+}
+
+function wrappedCwd() {
+  if (cachedCwd === '')
+    cachedCwd = rawMethods.cwd();
+  return cachedCwd;
 }
