+++ lib/internal/bootstrap/switches/is_main_thread.js
@@ -3,11 +3,6 @@
 const { ObjectDefineProperty } = primordials;
 const rawMethods = internalBinding('process_methods');
 
-process.abort = rawMethods.abort;
-process.umask = wrappedUmask;
-process.chdir = wrappedChdir;
-process.cwd = wrappedCwd;
-
 // TODO(joyeecheung): deprecate and remove these underscore methods
 process._debugProcess = rawMethods._debugProcess;
 process._debugEnd = rawMethods._debugEnd;
@@ -38,34 +33,6 @@ process.on('removeListener', stopListeningIfSignal);
 // ----              compare the setups side-by-side                    -----
 
 const { guessHandleType } = internalBinding('util');
-const {
-  parseMode,
-  validateString
-} = require('internal/validators');
-
-// Cache the working directory to prevent lots of lookups. If the working
-// directory is changed by `chdir`, it'll be updated.
-let cachedCwd = '';
-
-function wrappedChdir(directory) {
-  validateString(directory, 'directory');
-  rawMethods.chdir(directory);
-  // Mark cache that it requires an update.
-  cachedCwd = '';
-}
-
-function wrappedUmask(mask) {
-  if (mask !== undefined) {
-    mask = parseMode(mask, 'mask');
-  }
-  return rawMethods.umask(mask);
-}
-
-function wrappedCwd() {
-  if (cachedCwd === '')
-    cachedCwd = rawMethods.cwd();
-  return cachedCwd;
-}
 
 function createWritableStdioStream(fd) {
   let stream;
