+++ lib/internal/bootstrap/switches/is_not_main_thread.js
@@ -1,15 +1,6 @@
 'use strict';
 
 const { ObjectDefineProperty } = primordials;
-const rawMethods = internalBinding('process_methods');
-const {
-  unavailable
-} = require('internal/process/worker_thread_only');
-
-process.abort = unavailable('process.abort()');
-process.chdir = unavailable('process.chdir()');
-process.umask = wrappedUmask;
-process.cwd = rawMethods.cwd;
 
 delete process._debugProcess;
 delete process._debugEnd;
@@ -42,9 +33,6 @@ process.removeListener('removeListener', stopListeningIfSignal);
 const {
   createWorkerStdio
 } = require('internal/worker/io');
-const {
-  codes: { ERR_WORKER_UNSUPPORTED_OPERATION }
-} = require('internal/errors');
 
 let workerStdio;
 function lazyWorkerStdio() {
@@ -57,12 +45,3 @@ function getStdout() { return lazyWorkerStdio().stdout; }
 function getStderr() { return lazyWorkerStdio().stderr; }
 
 function getStdin() { return lazyWorkerStdio().stdin; }
-
-function wrappedUmask(mask) {
-  // process.umask() is a read-only operation in workers.
-  if (mask !== undefined) {
-    throw new ERR_WORKER_UNSUPPORTED_OPERATION('Setting process.umask()');
-  }
-
-  return rawMethods.umask(mask);
-}
